+++ src/Main.js
@@ -1,3 +1,4 @@
+import Layout from './pages/Layout';
 import Shared from './pages/Shared';
 import Folder from './pages/Folder';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
@@ -6,9 +7,10 @@ function Main() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Shared />} />
-        <Route path="/shared" element={<Shared />} />
-        <Route path="/folder" element={<Folder />} />
+        <Route path="/" element={<Layout />}>
+          <Route path="/shared" element={<Shared />} />
+          <Route path="/folder" element={<Folder />} />
+        </Route>
       </Routes>
     </BrowserRouter>
   );
