+++ src/components/Header/Header.js
@@ -1,6 +1,6 @@
-import logo from '../images/logo.svg';
-import '../style/header.css';
-import Profile from './Profile';
+import logo from '../../images/logo.svg';
+import './header.css';
+import Profile from '../Profile';
 import { Link } from 'react-router-dom';
 
 function Header() {
