+++ src/components/Shared/shared.css
