+++ src/pages/Layout.js
@@ -1,0 +1,15 @@
+import Header from '../components/Header/Header';
+import Footer from '../components/Footer/Footer';
+import { Outlet } from 'react-router-dom';
+
+function Layout() {
+  return (
+    <>
+      <Header />
+      <Outlet />
+      <Footer />
+    </>
+  );
+}
+
+export default Layout;
