+++ src/pages/Shared.js
@@ -1,7 +1,7 @@
 import { Helmet } from 'react-helmet';
-import SharedHeader from '../components/SharedHeader';
+import SharedHeader from '../components/Shared/SharedHeader';
 import CardList from '../components/CardList';
-import Footer from '../components/Footer';
+import Footer from '../components/Footer/Footer';
 
 function Shared() {
   return (
